% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Gehan.R
\name{Gehan}
\alias{Gehan}
\title{Gehan test}
\usage{
Gehan(x, y, alpha = 0.05, delta = 0.1, power = 0.9)
}
\arguments{
\item{x}{list of string.x is the Background Measurements.< can be put to specify less than the observed date.}

\item{y}{list of string.y is the Site Measurements. < can be put to specify less than the observed date.}

\item{alpha}{test significance for Gehan Test}

\item{delta}{the magnitude of the difference in median site and background that should be detected by Gehan Test}

\item{power}{Power value for Gehan Test}
}
\description{
Gehan test is used to do Gehan Test. Gehan test asks if the site median (MP/I) is larger than the background median (MB/G).
}
\examples{
for guidance example
y<-c("2","<4","8","17","20","25","34","<35","40","43")
x<-c("1","<4", "5", "7","<12","15","18","<21","<25","27")
Gehan(x,y,alpha=0.05,delta=2.0,power=0.9)
}

