% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RProjectSlip.R
\docType{data}
\name{Min_Num_Mea}
\alias{Min_Num_Mea}
\title{Min_Num_Mea
Before taking measurements for Slippage test you need to determine the proportion and power that you desire and see how many measurements you need
Min_Num_Mea is a matrix that calls a table from Gilbert and Simpson (1990, Table 1 and Figure 3) shows how many measurements needed}
\format{\preformatted{ num [1:10, 1:2] 60 40 30 25 15 15 15 10 10 10 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:10] ".10" ".15" ".20" ".25" ...
  ..$ : chr [1:2] ".80" ".90"
}}
\usage{
Min_Num_Mea
}
\description{
Min_Num_Mea
Before taking measurements for Slippage test you need to determine the proportion and power that you desire and see how many measurements you need
Min_Num_Mea is a matrix that calls a table from Gilbert and Simpson (1990, Table 1 and Figure 3) shows how many measurements needed
}
\keyword{datasets}

