% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QuantileTest.R
\name{QuanTest}
\alias{QuanTest}
\title{Quantile Test
The Quantile test looks at number of potentially impacted data among the largest data in the pooled basin-wide dataset.}
\usage{
QuanTest(x, y, Epsilon)
}
\arguments{
\item{x}{is the string of the Background Measurements.}

\item{y}{is string of the Site Measurements}

\item{Epsilon}{is the proportion of a site at which chemicals are present at concentrations greater than background levels.}
}
\description{
Quantile Test
The Quantile test looks at number of potentially impacted data among the largest data in the pooled basin-wide dataset.
}
\examples{
QuanTest(x,y,Epsilon)
set.seed(123)
Site <- floor(runif(30,1,100)) # input data for site measures
Bkg <- floor(runif(30,1,101))  # input data for background measures
tail((sort(c(Site,Bkg))),4)
QuanTest(Site,Bkg,.4)
}

