% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Gehan.R
\name{sampletest}
\alias{sampletest}
\title{Sample test for Gehan}
\usage{
sampletest(n, m, alpha = 0.05, delta = 2, power = 0.9)
}
\arguments{
\item{n}{list of string.x is the Background Measurements.}

\item{m}{list of string.y is the Site Measurements.}

\item{alpha}{test significance for Gehan Test}

\item{delta}{the magnitude of the difference in median site and background that should be detected by Gehan Test}

\item{power}{Power value for Gehan Test}
}
\description{
Sample test is used to judge whether there is enough sample for Gehan test
}
\examples{
for guidance example
y<-c("2","<4","8","17","20","25","34","<35","40","43")
x<-c("1","<4", "5", "7","<12","15","18","<21","<25","27")
sampletest(x,y,alpha=0.05,delta=2.0 ,power=0.9)
}

