% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RProjectSlip.R
\name{Slip}
\alias{Slip}
\title{Slippage Function
Slippage test compares measurements one the measurement of the background amount of chemcials and the other the new measurement to determine if the tested chemical is a chemicals of potential concern (COPC)}
\usage{
Slip(B, C, A, M, N)
}
\arguments{
\item{B}{is the string of the Background Measurements.}

\item{C}{is string of the Site Measurements}

\item{A}{is the desired Alpha .01 or .05 (Input either A01 or A05, respectively, in order to call appropriate critical value matrix)}

\item{M}{is the number of Background measurements taken}

\item{N}{is the number of New measurements taken to be compared to background measurements}
}
\description{
Slippage Function
Slippage test compares measurements one the measurement of the background amount of chemcials and the other the new measurement to determine if the tested chemical is a chemicals of potential concern (COPC)
}
\examples{
Slip("Background Measurement", "New Measurements", "A01 or A05", #, #)
B<-c(23, 36, 37, 37, 44, 57, 60, 61, 61, 79)
C<-c(15, 15, 20, 29, 30, 39, 60, 89, 90, 100)
Slip(B, C, A05, 10, 10)
}

